Move django files and dirs into separate dir
Once we get a bunch of django apps; is easier to go to that dir
and do a search in 'Django only' files. Or go outside of dir
and do only a one level deep search in non-django files.

Again something to discuss with other engineers first as the argument
can be made that not implementing this change would result in less
directory hopping.

Imo the total time saved by the search convenience + clarity in
navigation, especially from the moment the project grows to a
considerable size, would be greater than the time needed to not have to do an
extra hop (if you use a tool like tmux it's straightforward to
spawn multiple windows from the current 'hopped' directory).